+++ lib/presentation/pages/exercise/dashboard/widgets/add_button.dart
@@ -1,28 +1,16 @@
 import 'package:flutter/material.dart';
 
-import '../../../../../theme/pp_light_theme.dart';
-import '../../../../router/route_paths.dart';
+import 'package:power_progress/presentation/widgets/main_button.dart';
+import 'package:power_progress/presentation/router/route_paths.dart';
 
 class AddButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return FloatingActionButton(
+    return MainButton(
       onPressed: () {
         Navigator.of(context).pushNamed(RoutePaths.exerciseAdd);
       },
-      child: Ink(
-        decoration: BoxDecoration(
-          gradient: PPTheme.pinkYellowGradient,
-          borderRadius: BorderRadius.circular(30.0),
-        ),
-        child: Container(
-          constraints: const BoxConstraints(maxWidth: 300.0, minHeight: 50.0),
-          alignment: Alignment.center,
-          child: const Icon(
-            Icons.add,
-          ),
-        ),
-      ),
+      icon: Icons.add,
     );
   }
 }